Add cargo package manager support for NetBSD
- This commit also fixes the many compilation errors that occur on
  NetBSD
- Move count_cargo() to shared directory
- Small tweaks to macOS documentation
