+++ Cargo.toml
@@ -15,7 +15,6 @@ cfg-if = "1.0.0"
 libc = "0.2.88"
 local_ipaddress = "0.1.3"
 lazy_static = "1.4.0"
-itertools = "0.10.0"
 
 [target.'cfg(target_os = "netbsd")'.dependencies]
 nix = "0.20.0"
@@ -23,10 +22,11 @@ regex = "1.4.5"
 
 [target.'cfg(any(target_os = "linux", target_os = "netbsd"))'.dependencies]
 os-release = "0.1"
+itertools = "0.10.0"
+home = "0.5.3"
 
 [target.'cfg(target_os = "linux")'.dependencies]
 sqlite = "0.25.3"
-home = "0.5.3"
 
 [target.'cfg(target_os = "macos")'.dependencies]
 core-foundation = "0.9.1"
